@@ -1,62 +1,55 @@
+// imports de dependências de funcionalidades nativas do angular
 import { BrowserModule } from '@angular/platform-browser';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { HttpClientModule } from '@angular/common/http';
+import { LocationStrategy , HashLocationStrategy } from '@angular/common';
 import { NgModule, LOCALE_ID } from '@angular/core';
-import { HttpModule } from '@angular/http';
-import { RouterModule } from '@angular/router';
-import { FormsModule } from '@angular/forms';
+import { RouterModule, PreloadAllModules } from '@angular/router';
 
+// import de dependências de rotas
 import { ROUTES } from './app.routes';
 
+// imports de módulos
+import { SharedModule } from './shared/shared.module';
+
+// imports de dependências de componentes
 import { AppComponent } from './app.component';
-import { AboutComponent } from './about/about.component';
 import { HeaderComponent } from './header/header.component';
 import { HomeComponent } from './home/home.component';
-import { InputComponent } from './shared/input/input.component';
 import { MenuComponent } from './restaurant-detail/menu/menu.component';
 import { MenuItemComponent } from './restaurant-detail/menu-item/menu-item.component';
-import { OrderComponent } from './order/order.component';
-import { OrderItemsComponent } from './order/order-items/order-items.component';
-import { RadioComponent } from './shared/radio/radio.component';
+import { OrderSummaryComponent } from './order-summary/order-summary.component';
 import { RestaurantComponent } from './restaurants/restaurant/restaurant.component';
 import { RestaurantDetailComponent } from './restaurant-detail/restaurant-detail.component';
 import { RestaurantsComponent } from './restaurants/restaurants.component';
 import { ReviewsComponent } from './restaurant-detail/reviews/reviews.component';
 import { ShoppingCartComponent } from './restaurant-detail/shopping-cart/shopping-cart.component';
-
-import { OrderService } from './order/order.service';
-import { RestaurantsService } from './restaurants/restaurants.service';
-import { ShoppingCartService } from './restaurant-detail/shopping-cart/shopping-cart.service';
-import { DeliveryCostsComponent } from './order/delivery-costs/delivery-costs.component';
-import { OrderSummaryComponent } from './order-summary/order-summary.component';
-import { RatingComponent } from './shared/rating/rating.component';
+import { NotFoundComponent } from './not-found/not-found.component';
 
 @NgModule({
   declarations: [
     AppComponent,
     HeaderComponent,
     HomeComponent,
-    AboutComponent,
+    MenuComponent,
+    MenuItemComponent,
+    OrderSummaryComponent,
     RestaurantsComponent,
     RestaurantComponent,
     RestaurantDetailComponent,
-    MenuComponent,
-    ShoppingCartComponent,
-    MenuItemComponent,
     ReviewsComponent,
-    OrderComponent,
-    InputComponent,
-    RadioComponent,
-    OrderItemsComponent,
-    DeliveryCostsComponent,
-    OrderSummaryComponent,
-    RatingComponent
+    ShoppingCartComponent,
+    NotFoundComponent
   ],
   imports: [
     BrowserModule,
-    HttpModule,
-    FormsModule,
-    RouterModule.forRoot(ROUTES)
+    BrowserAnimationsModule,
+    HttpClientModule,
+    RouterModule.forRoot(ROUTES, {preloadingStrategy: PreloadAllModules}),
+    SharedModule.forRoot(),
   ],
-  providers: [RestaurantsService, ShoppingCartService, OrderService, {provide: LOCALE_ID, useValue: 'pt-BR'}],
+  providers: [ /*{provide: LocationStrategy, useClass: HashLocationStrategy},*/ {provide: LOCALE_ID, useValue: 'pt-BR'} ],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
