@@ -1,13 +1,13 @@
 import { Routes } from '@angular/router';
 
 import { HomeComponent } from './home/home.component';
-import { AboutComponent } from './about/about.component';
+import { MenuComponent } from './restaurant-detail/menu/menu.component';
+import { OrderSummaryComponent } from './order-summary/order-summary.component';
 import { RestaurantsComponent } from './restaurants/restaurants.component';
 import { RestaurantDetailComponent } from './restaurant-detail/restaurant-detail.component';
-import { MenuComponent } from './restaurant-detail/menu/menu.component';
 import { ReviewsComponent } from './restaurant-detail/reviews/reviews.component';
-import { OrderComponent } from './order/order.component';
-import { OrderSummaryComponent } from './order-summary/order-summary.component';
+import { NotFoundComponent } from './not-found/not-found.component';
+
 
 export const ROUTES: Routes = [
     { path: '', component: HomeComponent },
@@ -18,7 +18,8 @@ export const ROUTES: Routes = [
             { path: 'menu', component: MenuComponent },
             { path: 'reviews', component: ReviewsComponent }
         ]},
-    { path: 'order', component: OrderComponent },
+    { path: 'order', loadChildren: './order/order.module#OrderModule' },
     { path: 'order-summary', component: OrderSummaryComponent },
-    { path: 'about', component: AboutComponent }
+    { path: 'about', loadChildren: './about/about.module#AboutModule' },
+    { path: '**', component: NotFoundComponent } // ** --> rota de wildcard
 ];
