@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { FormGroup, FormBuilder, Validators, AbstractControl } from '@angular/forms';
 import { Router } from '@angular/router';
 
 import { CartItem } from '../restaurant-detail/shopping-cart/cart-item.model';
@@ -12,6 +13,9 @@ import { RadioOption } from '../shared/radio/radio-option.model';
 })
 export class OrderComponent implements OnInit {
 
+  emailPattern = /^(([^<>()\[\]\.,;:\s@\"]+(\.[^<>()\[\]\.,;:\s@\"]+)*)|(\".+\"))@(([^<>()[\]\.,;:\s@\"]+\.)+[^<>()[\]\.,;:\s@\"]{2,})$/;
+  numberPattern = /^[0-9]*$/;
+  orderForm: FormGroup;
   delivery = 8;
 
   paymentOptions: RadioOption[] = [
@@ -20,9 +24,33 @@ export class OrderComponent implements OnInit {
     { label: 'Cartão Refeição', value: 'REF' }
   ];
 
-  constructor(private orderService: OrderService, private router: Router) { }
+  static equalsTo(group: AbstractControl): { [key: string]: boolean } {
+    const email = group.get('email');
+    const emailConfirmation = group.get('emailConfirmation');
+    if (!email || !emailConfirmation) {
+      return undefined;
+    }
+
+    if (email.value !== emailConfirmation.value) {
+      return { emailsNotMatch: true };
+    }
+    return undefined;
+  };
+
+  constructor(private orderService: OrderService,
+              private router: Router,
+              private formBuilder: FormBuilder) { }
 
   ngOnInit() {
+    this.orderForm = this.formBuilder.group({
+      name: this.formBuilder.control('', [Validators.required, Validators.minLength(4)]),
+      email: this.formBuilder.control('', [Validators.required, Validators.pattern(this.emailPattern)] ),
+      emailConfirmation: this.formBuilder.control('', [Validators.required, Validators.pattern(this.emailPattern)]),
+      address: this.formBuilder.control('', [Validators.required, Validators.minLength(4)]),
+      number: this.formBuilder.control('', [Validators.required, Validators.pattern(this.numberPattern)]),
+      optionalAddress: this.formBuilder.control(''),
+      paymentOption: this.formBuilder.control('', [Validators.required])
+    }, {validator: OrderComponent.equalsTo});
   }
 
   itemsValue(): number {
