@@ -1,21 +1,19 @@
+import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Http, Headers, RequestOptions } from '@angular/http';
 
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/operator/map';
 
-import { MEAT_API } from '../app.api';
-
-import { ShoppingCartService } from '../restaurant-detail/shopping-cart/shopping-cart.service';
 import { CartItem } from '../restaurant-detail/shopping-cart/cart-item.model';
+import { MEAT_API } from '../app.api';
 import { Order, OrderItem } from './order.model';
-
+import { ShoppingCartService } from '../restaurant-detail/shopping-cart/shopping-cart.service';
 
 
 @Injectable()
 export class OrderService {
 
-    constructor(private cartService: ShoppingCartService, private http: Http) {}
+    constructor(private cartService: ShoppingCartService, private http: HttpClient) {}
 
     itemsValue(): number {
         return this.cartService.total();
@@ -43,12 +41,7 @@ export class OrderService {
 
     // No Observable abaixo, poderia trocar o tipo string para Order e retirar o segundo map do método.
     checkOrder(order: Order): Observable<string> {
-        const headers = new Headers();
-        headers.append('Content-Type', 'application/json');
-        return this.http.post(`${MEAT_API}/orders`,
-                               JSON.stringify(order),
-                               new RequestOptions({headers: headers}))
-                        .map(response => response.json())
+        return this.http.post<Order>(`${MEAT_API}/orders`, order)
                         .map(order => order.id);
     };
 
