@@ -6,6 +6,6 @@ export interface Restaurant {
     deliveryEstimate: string;
     rating: number;
     imagePath: string;
-    about: string;
-    hours: string;
+    about?: string;
+    hours?: string;
 };
