@@ -1,5 +1,5 @@
 import { Component, OnInit, Input, ContentChild, AfterContentInit } from '@angular/core';
-import { NgModel } from '@angular/forms';
+import { NgModel, FormControlName } from '@angular/forms';
 
 @Component({
   selector: 'mt-input-container',
@@ -13,6 +13,7 @@ export class InputComponent implements OnInit, AfterContentInit {
   input: any;
 
   @ContentChild(NgModel) model: NgModel;
+  @ContentChild(FormControlName) control: FormControlName;
 
   constructor() { }
 
@@ -20,9 +21,9 @@ export class InputComponent implements OnInit, AfterContentInit {
   }
 
   ngAfterContentInit() {
-    this.input = this.model;
+    this.input = this.model || this.control;
     if (this.input === undefined) {
-      throw new Error ('Esse componente precisa ser usado com uma diretiva ngModel');
+      throw new Error ('Esse componente precisa ser usado com uma diretiva ngModel ou FormControlName.');
     }
   }
 
